Convert from _countof to size

+++ lib/dim/include/dim/task.h
@@ -21,14 +21,14 @@ private:
 };
 
 void DimTaskPushEvent (IDimTaskNotify & task);
-void DimTaskPushEvent (IDimTaskNotify * tasks[], int numTasks);
+void DimTaskPushEvent (IDimTaskNotify * tasks[], size_t numTasks);
 
 void DimTaskPushCompute (IDimTaskNotify & task);
-void DimTaskPushCompute (IDimTaskNotify * tasks[], int numTasks);
+void DimTaskPushCompute (IDimTaskNotify * tasks[], size_t numTasks);
 
-HDimTaskQueue DimTaskCreateQueue (const std::string & name, unsigned threads);
-void DimTaskSetQueueThreads (HDimTaskQueue q, unsigned threads);
+HDimTaskQueue DimTaskCreateQueue (const std::string & name, int threads);
+void DimTaskSetQueueThreads (HDimTaskQueue q, int threads);
 void DimTaskPush (HDimTaskQueue q, IDimTaskNotify & task);
-void DimTaskPush (HDimTaskQueue q, IDimTaskNotify * tasks[], int numTasks);
+void DimTaskPush (HDimTaskQueue q, IDimTaskNotify * tasks[], size_t numTasks);
 
 #endif
